Refactor: Validation, Game class 분리
