+++ src/main/java/validation/Validation.java
@@ -1,0 +1,19 @@
+package validation;
+
+public class Validation {
+    public static final int MIN_RANDOM_NUMBER = 1;
+    public static final int MAX_RANDOM_NUMBER = 9;
+    public static final int RANDOM_NUMBER_SIZE = 3;
+
+    public static String START_STRING = "숫자 야구 게임을 시작합니다.";
+    public static String INPUT_STRING = "숫자를 입력해주세요 : ";
+    public static String WRONG_INPUT_SIZE = "입력 문자열의 길이가 3이 아닙니다.";
+    public static String WRONG_INPUT_TYPE = "숫자가 아닌 문자가 포함되어 있습니다.";
+    public static String NOTHING = "낫싱";
+    public static String END_GAME_STRING = "3개의 숫자를 모두 맞히셨습니다! 게임 종료";
+    public static String RESTART_STRING = "게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.";
+    public static String WRONG_INPUT_STRING = "restart() error";
+    Validation() {
+
+    }
+}
